requirements.txt
torch==2.9.1
torchvision==0.24.1
scikit-learn==1.8.0
PyYAML==6.0.3
numpy==2.4.1